+++ Example/PinLayoutSample/Supporting Files/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -128,6 +128,12 @@
       "idiom" : "ipad",
       "filename" : "Icon-App-83.5x83.5@2x.png",
       "scale" : "2x"
+    },
+    {
+      "size" : "1024x1024",
+      "idiom" : "ios-marketing",
+      "filename" : "Icon-App-1024.png",
+      "scale" : "1x"
     }
   ],
   "info" : {
